Final tests for ObjectConfig
report_unused_values was tricky, and this implementation isn't super elegant. Note that there is an issue with this method in that it assumes messages is a list of strings, and will therefore log each character individually if it is not a list. The test does not account for this error, and therefore currently does not pass.

describe_unused_values is almost an abstract method, but has to have an implementation for report_unused_values and therefore is simply tested by asserting it returns an empty list.