Next Object, Structure Improvements
It makes sense to combine ObjectConfig's tests with tests for its children, so in order to do that I made a utility class (in this file for now, can be broken out later) to reuse assert_eq, and other custom assertions without duplicating code.

ListConfig tests will mirror ObjectConfig's, with additional considerations. I test the init in the same way, with an added assertion that a ListConfig object is actually a child of ObjectConfig - the inherited properties from the parent are not tested to avoid duplicaiton.